% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmcparse.R
\docType{package}
\name{cmcparse}
\alias{cmcparse}
\alias{cmcparse-package}
\title{cmcparse: A package for parse CMC Markets report from exported CSV file.}
\description{
The cmcparse package provides three categories of important functions:
readCMCReport, performanceCMCreport and writeCMCreportToCsv.
}
\section{cmcparse functions}{

The foo functions...
}
\author{
Mario Pisa
}
\seealso{
\url{https://www.cmcmarkets.com/}
}

