% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceCMCreport.R
\name{performanceCMCreport}
\alias{performanceCMCreport}
\title{Operations performance from data frame generated by readCMCreport.R}
\usage{
performanceCMCreport(df)
}
\arguments{
\item{df}{data.frame generated by readCMCreport.R}
}
\value{
screen print lines
}
\description{
\code{performanceCMCreport} returns a print lines with statistic data and
plot key indicators.
}
\examples{
performanceCMCreport(operations)
performanceCMCreport(readCMCreport("../myCMCreports/Historico20160909.csv"))
}
\author{
Mario Pisa
}
\seealso{
\url{https://www.cmcmarkets.com/}
}

