% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCMCreport.R
\name{readCMCreport}
\alias{readCMCreport}
\title{Parse CMC Markets report from exported CSV file.}
\usage{
readCMCreport(file)
}
\arguments{
\item{file}{report exported from CMC Market as csv file}
}
\value{
data.frame of operations one by one and added useful info and
   calculations.
}
\description{
\code{readCMCreport} returns a data.frame with the operations one by one and
to add useful information and statistics calculations.
}
\examples{
\dontrun{readCMCreport("../myCMCreports/Historico20160909.csv")}
\dontrun{operations <- readCMCreport("../myCMCreports/Historico20160909.csv")}
}
\author{
Mario Pisa
}
\seealso{
\url{https://www.cmcmarkets.com/}
}

