% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeCMCreportToCsv.R
\name{writeCMCreportToCsv}
\alias{writeCMCreportToCsv}
\title{Write operations into csv file from data frame generated by readCMCreport.R}
\usage{
writeCMCreportToCsv(df)
}
\arguments{
\item{df}{data.frame generated by readCMCreport.R}
}
\value{
csv file
}
\description{
\code{writeCMCreportToCsv} returns a csv file with one operation per row.
}
\examples{
\dontrun{writeCMCreportToCsv(operations)}
}
\author{
Mario Pisa
}
\seealso{
\url{https://www.cmcmarkets.com/}
}

